pyproject.toml
[project]
name = "command-lab-official-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ap12-code", email = "ap12@ap12.net" }
]
dependencies = [
    "discord>=2.3.2",
    "pydantic>=2.9.2",
    "jishaku>=2.5.2",
    "markdownify>=0.13.1",
    "aiohttp>=3.10.8",
    "aiofiles>=24.1.0",
    "tqdm>=4.66.5",
    "pygit2>=1.15.1",
    "pyyaml>=6.0.2",
    "lorem>=0.1.1",
    "pillow>=10.4.0",
    "nbtlib>=2.0.4",
    "multipledispatch>=1.0.0",
    "pyrr>=0.10.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pyproject-flake8>=7.0.0",
    "isort>=5.13.2",
    "black>=24.8.0",
    "readchar>=4.2.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/command_lab_official_bot"]

[tool.isort]
profile = "black"
line_length = 119

[tool.flake8]
extend-ignore = "E203"
max-line-length = 119

[tool.black]
line-length = 119
